+++ src/db/db_mysql/mod.rs
@@ -483,6 +483,8 @@ impl DbConnection for MysqlConnection {
             .filter_map(|station| Url::parse(&station.url_resolved).ok())
             .map(|mut url| {
                 url.set_path("/");
+                url.set_query(None);
+                url.set_fragment(None);
                 url.to_string()
             })
             .collect();