+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* DB: removed fragment and query from streaming server urls
+* CHECK: removed fragment and query from streaming server urls
+
 ## [0.7.17] 2021-10-10
 ### Added
 * API: new order type "changetimestamp"